Add files via upload
This version of Extract Machine utilizes PyMuPDF to grab text from a specified coordinate range. Use the first cell to draw a box based on the coordinates that you provide. Once you verify the coordinates encompass the right boxes you want to grab data, run the extract machine. Remember to enter the right file location and the path to where you would like the file to be extracted